third_party/blink/web_tests/fast/forms/suggestion-picker/week-suggestion-picker-mouse-operations-expected.txt

Check that page popup doesn't exist at first.
PASS $("mock-page-popup") is null
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that hovering over an entry highlights it.
PASS highlightedEntry() is "2012-W02"
Check that moving the mouse outside the popup de-highlights entries.
PASS highlightedEntry() is null
Check that mouse click closes the popup and sets the value.
PASS $("mock-page-popup") is null
PASS document.getElementById("week").value is "2012-W03"
Check that you can scroll using the mouse wheel.
PASS scrollTopBeforeWheelEvent < suggestionList.scrollTop became true
PASS scrollTopBeforeWheelEvent > suggestionList.scrollTop became true
PASS $("mock-page-popup") is null
Check that you can scroll using the scroll bar.
PASS scrollTopBeforeWheelEvent < scrollTopAfterWheelEvent is true
PASS successfullyParsed is true

TEST COMPLETE
